Move Http error cause loggers outside of api module.
