+++ components/proxy/src/test/java/com/hotels/styx/proxy/HttpErrorStatusCauseLoggerTest.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.metrics;
+package com.hotels.styx.proxy;
 
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.support.matchers.LoggingTestSupport;
