+++ components/server/src/main/java/com/hotels/styx/server/RequestProgressListener.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.metrics;
+package com.hotels.styx.server;
 
 /**
  * A listener interface for progression of HTTP requests.
